Move process_assignment and find_if_eq_diseq to solver_state.rs
These solver-level functions are now free functions in solver_state.rs
instead of congruence_closure.rs. They remain free functions (not
methods on SolverState) because Egraph::union_process_assignment still
calls process_assignment directly — this will be removed once
watch-based propagation is implemented.

congruence_closure.rs now only contains pure ProofForestEdge helpers:
add_parent, get_parent, get_child.
